Moves event handlers out of app.
App now just contains app composition.
