+++ src/lib.rs
@@ -1,2 +1,4 @@
 pub mod app;
+pub mod event_handler;
 pub mod intro;
+pub mod voice_handler;
